+++ src/main/java/com/teamtea/eclipticseasons/compat/jei/fake/GreenHouseCoreRecipeCategory.java
@@ -3,6 +3,7 @@ package com.teamtea.eclipticseasons.compat.jei.fake;
 import com.teamtea.eclipticseasons.api.EclipticSeasonsApi;
 import com.teamtea.eclipticseasons.api.constant.solar.SolarTerm;
 import com.teamtea.eclipticseasons.api.util.SimpleUtil;
+import com.teamtea.eclipticseasons.common.block.GreenHouseCoreBlock;
 import com.teamtea.eclipticseasons.common.registry.ItemRegistry;
 import com.teamtea.eclipticseasons.compat.jei.ESJEIPlugin;
 import com.teamtea.eclipticseasons.config.CommonConfig;
@@ -21,6 +22,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.FormattedCharSequence;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.crafting.Ingredient;
 import org.jetbrains.annotations.NotNull;
 
@@ -67,6 +69,23 @@ public class GreenHouseCoreRecipeCategory implements IRecipeCategory<GreenHouseC
 
         builder.addSlot(RecipeIngredientRole.INPUT,33+8,3)
                 .addIngredients(Ingredient.of(recipe.input()));
+
+
+
+        if (recipe.end() instanceof GreenHouseCoreBlock greenHouseCoreBlock) {
+            Item essenceItem = switch (greenHouseCoreBlock.getSeason()) {
+                case SPRING -> ItemRegistry.spring_greenhouse_essence_item.get();
+                case SUMMER -> ItemRegistry.summer_greenhouse_essence_item.get();
+                case AUTUMN -> ItemRegistry.autumn_greenhouse_essence_item.get();
+                case WINTER -> ItemRegistry.winter_greenhouse_essence_item.get();
+                default -> null;
+            };
+
+            if (essenceItem != null) {
+                builder.addInvisibleIngredients(RecipeIngredientRole.OUTPUT)
+                        .addItemLike(essenceItem);
+            }
+        }
     }